Chaotic analysis is now possible*
The jacobian and inplace 1D n-forest rule are
defined and can be used in functions like
lyapunovspectrum. The system does not appear
to be a periodic system, but one might quantify
it's chaotic dynamics now... look into this
and consider what other analysis could be
done or what other simulations could be done...
write a script that basically does the same
as dominiques simulation though consider
what other parameters might be worth varying.
Can a power law be extracted somewhere???
It seems like increasing the number of systems
is not really worth it because the systems
tend toward the same dynamics, at least
in the biotic pumping mechanism, but
quantifying the chaos of this mechanism
would still likely be useful right?

 Please enter the commit message for your changes. Lines starting
 with '' will be ignored, and an empty message aborts the commit.

 On branch main
 Your branch is ahead of 'origin/main' by 2 commits.
   (use "git push" to publish your local commits)

 Changes to be committed:
	modified:   src/chaos.jl